combines prompts and resources components, removes unneeded code, edits navbar routes
