+++ src/Pages/LandingPage.css
@@ -23,8 +23,7 @@
   margin-bottom: 1rem;
 }
 
-.content-wrapper p {
-  color: #5372FF; 
+.content-wrapper p { 
   line-height: 1.6;
   font-size: 1rem;
   margin-bottom: 2rem;
