+++ src/Pages/LandingPage.jsx
@@ -18,9 +18,8 @@ const LandingPage = () => {
     <div className="landing-page">
       <section className="intro-section">
         <div className="content-wrapper">
-          <h1>Our aim is to elevate your interview skills to new heights</h1>
-          <p>Our community-driven app allows users to engage in mock interviews with fellow members, fostering a supportive environment for skill honing. Earn badges and climb the ranks as you showcase your interviewing prowess. Stay tuned for exciting updates, including the integration of AI technology!</p>
-          <button className="button contact-button">Contact Us</button>
+          <h1>We'll elevate your interviewing skills to new heights!</h1>
+          <p>Our community-driven app allows users to engage in mock interviews with professional industry leaders, fostering a supportive environment for skill honing. Earn badges and climb the ranks as you showcase your interviewing prowess. Stay tuned for exciting updates, including the integration of AI technology!</p>
         </div>
         <div className="image-wrapper">
           <img src="src/assets/interview pic landing page.jpeg" alt="Group studying" />
